Renamed InitialLoggerContext to LoggerContextRule.
